+++ packages/gcode-containers/src/index.ts
@@ -18,5 +18,5 @@ export {
   streamEntry
 } from './zip.js';
 export type { ContainerLimits, ZipEntry, ZipDirectory } from './zip.js';
-export { openGcode3mf, sniffGcode3mf } from './gcode-3mf.js';
+export { openGcode3mf, sniffGcode3mf, filamentColoursFromSettings } from './gcode-3mf.js';
 export type { OpenedContainer, PlateEntry, ContainerMetadata, StreamLike } from './gcode-3mf.js';