+++ packages/gcode-renderer-2d/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @chestnutlabs/gcode-renderer-2d
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chestnutlabs/gcode-colors@0.8.1
+  - @chestnutlabs/toolpath-core@0.8.1
+
+## 0.8.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chestnutlabs/gcode-colors@0.8.0
+  - @chestnutlabs/toolpath-core@0.8.0
+
 ## 0.7.0
 
 ### Patch Changes